Create and check a 'last-updated' file for up-to-dateness of mirrors, as opposed to 'repomd.xml' which may not get updated if the repo doesn't have any new RPMs
The recent change to skip running createrepo if the packages haven't
changed between runs causes mirror file generation to fail.  This is
because the test for whether a mirror is up to date is based on how old
the repomd.xml (which createrepo creates) is.  We give mirrors one day
before marking them outdated.

We're supposed to always have a fallback mirror (the repo server itself)
but we perform the same check for its up-to-dateness.  If createrepo has
been skipped for long enough, then even the main repo server was
considered out of date with the main repo server.

Instead, create sentinel 'last-updated' files that get updated whether
createrepo is run or skipped; use that instead of repomd.xml.

Along the way, I dropped the "has the repo migrated from the pre-OSG-23
layout" check because it's out of date.